"Refactoring utility functions into individual classes"
In this commit, utility functions previously organized in one class were refactored into separate classes grouped by their functionality to improve code management and readability. 'Util.cs' functions were split into three new classes: 'ProcessUtils.cs', 'FileUtils.cs', 'LogUtils.cs'. Changes were applied to all affected files that were previously using methods from the 'Util.cs' class.
