+++ FoliCon/App.xaml.cs
@@ -1,4 +1,5 @@
-﻿using FoliCon.ViewModels;
+﻿using FoliCon.Modules.utils;
+using FoliCon.ViewModels;
 using NLog;
 using Prism.Ioc;
 using Sentry;
@@ -20,7 +21,7 @@ public partial class App
     public App()
     {
         
-        LogManager.Configuration = Util.GetNLogConfig();
+        LogManager.Configuration = LogUtils.GetNLogConfig();
         DispatcherUnhandledException += App_DispatcherUnhandledException;
         GlobalDataHelper.Load<AppConfig>();
         Logger.Info("FoliCon Initilized");
